when upgrading, new upgrade stats updated in tower
upgradeStats object attatched in objects in path array to tell what stats to upgrade. This object will be looped through and have its named be compared to a tower's stats and added to it. That stat will be the tower's new stat and the tower's method will use those values to do everything