reuse tower arr indexes
when tower sold or somehow killed, that tower is still kept in the arr. When the maximum amount of towers in the arr is reached, a tower that has been killed will be replaced with a new tower. I dont trust sprite.destroy(), so hopefully this will put the sprite and its class up for garbage collection