enemy motion looks snakey
sprite's x value is given a random value to be added or subtracted from its current x value. Every 100 times this happens, a new random value is selected. After 50 movements, the direction is changed. All while this is happening, the enemy is moving in the y direction as well, giving it an overall snakey-s movement path